update and test correct arguments
